add icons
