+++ components/form-builder/icons/GlobeIcon.tsx
@@ -1,0 +1,19 @@
+import React from "react";
+export const GlobeIcon = ({ className, title }: { className?: string; title?: string }) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    height="34"
+    width="34"
+    className={className}
+    viewBox="0 0 34 34"
+    focusable="false"
+    aria-hidden={title ? false : true}
+    role={title ? "img" : "presentation"}
+  >
+    {title && <title>{title}</title>}
+    <path
+      d="M17 33.6668C14.6945 33.6668 12.5278 33.2293 10.5 32.3543C8.47226 31.4793 6.70837 30.2918 5.20837 28.7918C3.70837 27.2918 2.52087 25.5279 1.64587 23.5002C0.770874 21.4724 0.333374 19.3057 0.333374 17.0002C0.333374 14.6946 0.770874 12.5279 1.64587 10.5002C2.52087 8.47239 3.70837 6.7085 5.20837 5.2085C6.70837 3.7085 8.47226 2.521 10.5 1.646C12.5278 0.770996 14.6945 0.333496 17 0.333496C19.3056 0.333496 21.4723 0.770996 23.5 1.646C25.5278 2.521 27.2917 3.7085 28.7917 5.2085C30.2917 6.7085 31.4792 8.47239 32.3542 10.5002C33.2292 12.5279 33.6667 14.6946 33.6667 17.0002C33.6667 19.3057 33.2292 21.4724 32.3542 23.5002C31.4792 25.5279 30.2917 27.2918 28.7917 28.7918C27.2917 30.2918 25.5278 31.4793 23.5 32.3543C21.4723 33.2293 19.3056 33.6668 17 33.6668ZM15.2084 31.1252V27.7085C14.2362 27.7085 13.4167 27.3474 12.75 26.6252C12.0834 25.9029 11.75 25.0557 11.75 24.0835V22.2502L3.20837 13.7085C3.06948 14.2641 2.97226 14.8127 2.91671 15.3543C2.86115 15.896 2.83337 16.4446 2.83337 17.0002C2.83337 20.6113 4.00699 23.764 6.35421 26.4585C8.70143 29.1529 11.6528 30.7085 15.2084 31.1252ZM27.4584 26.6252C28.0695 25.9585 28.6042 25.2502 29.0625 24.5002C29.5209 23.7502 29.9098 22.9654 30.2292 22.146C30.5487 21.3266 30.7848 20.4863 30.9375 19.6252C31.0903 18.7641 31.1667 17.8891 31.1667 17.0002C31.1667 14.0557 30.3611 11.3821 28.75 8.97933C27.1389 6.57655 24.9862 4.81961 22.2917 3.7085V4.4585C22.2917 5.43072 21.9584 6.27794 21.2917 7.00016C20.625 7.72238 19.8056 8.0835 18.8334 8.0835H15.2084V11.7085C15.2084 12.1807 15.0209 12.5696 14.6459 12.8752C14.2709 13.1807 13.8473 13.3335 13.375 13.3335H9.91671V17.0002H20.6667C21.1389 17.0002 21.5278 17.1807 21.8334 17.5418C22.1389 17.9029 22.2917 18.3196 22.2917 18.7918V24.0835H24.0834C24.8889 24.0835 25.5973 24.3196 26.2084 24.7918C26.8195 25.264 27.2362 25.8752 27.4584 26.6252Z"
+      fill="black"
+    />
+  </svg>
+);
